requirements.txt
beautifulsoup4
requests
pdfkit
mysql-connector-python
tk